Added a helper method to convert our representation of an amount to a string with 2 decimal points. Made small changes in the gui, it will now start with the first portfolio selected.

+++ stockrobot/src/database/jpa/tables/PortfolioEntitys.java
@@ -39,7 +39,7 @@ public class PortfolioEntitys {
 	private AlgorithmEntitys algorithm;
 	
 	@Column
-	private long balance;
+	private Long balance;
 	
 	@Column
 	private boolean watchAllStocks;
@@ -67,7 +67,7 @@ public class PortfolioEntitys {
 	public PortfolioEntitys(String name) {
 		this.name = name;
 		algorithm = null;
-		balance = 0;
+		balance = (long)0;
 		watchAllStocks = false;
 	}
 	/**
